Fix tokens defined by reference to environment variables
Users of private npm modules (especialy in conjunction with npm organizations) commonly have package-local `.npmrc` files that use an environment variable as a placeholder for the auth token. The npm cli dereferences these placeholders automatically. It would be preferable if this module matched that behavior.

Currently, this lack of parity impacts the wombat tool. npm/wombat-cli#11