fix(auth): legacy token is basic auth
Change the auth type for the legacy `_auth` token to `Basic`.

`_auth` is just the Base64 encoded value of `username:password`.

It should be passed as-is in a `Basic` HTTP(S) authentication header.

Any attempt to send the value of `_auth` as a `Bearer` token will cause
npm to fail to authenticate, leading to a 401 HTTP error.

Reference regarding `_auth` - https://github.com/npm/npm/issues/6162#issuecomment-55281038